requirements.txt
# Travel Agent - Backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
google-adk>=1.18.0
pydantic>=2.0
python-dotenv>=1.0.0
httpx>=0.27.0
python-multipart>=0.0.9
